use string_view instead of const char* in PerformanceObserver APIs (#39513)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39513

changelog: [internal]

Using std::hash with const char * is dangerous, as it will not hash the contents but only the pointer itself. We want to adopt std::hash in favour of folly to reduce our dependency on folly. One way to fix the issue with const char * is to use std::string_view.

Reviewed By: rshest, cipolleschi

Differential Revision: D49355536

fbshipit-source-id: fd2a7b4558461e913cdf0ec0e8fe878cae2f6015