Docs update: PixelRatio.getFontScale (#39709)
Summary:
Addresses this issue: https://github.com/facebook/react-native/issues/39708

The embedded docs for `PixelRatio.getFontScale` haven't been updated in a while and it is still claiming lack of support for iOS when this is not the reality anymore. This was confusing for me when looking for docs through my local workspace.

## Changelog:

[INTERNAL] [CHANGED] - Updated PixelRatio.getFontScale docs to match the latest website

Pull Request resolved: https://github.com/facebook/react-native/pull/39709

Test Plan: N/A

Reviewed By: rshest

Differential Revision: D49752103

Pulled By: NickGerleman

fbshipit-source-id: 1b220bc0210639d7863b6961f665aab2cba52889