Remove JS-side Platform checks For Android <= 23 (#39771)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39771

We now support SDK 23 as a minimum. That means we can drop JS side checks.

Because Android has `Platform.Version` as a number, and iOS as a string, it's common to do an Android check before checking version, even if in Android only code. I tried to remove those where it was super clear the code was Android only (i.e. the file is `.android.js`), but otherwise leave the Android platform check to not give the possibility of changing behavior.

Changelog:
[Internal]

Reviewed By: luluwu2032

Differential Revision: D49814610

fbshipit-source-id: f28b09db7091598e187fee0f383561e1c1993e9a