Reduce visibility of ExecutorException
Summary:
In an attempt to reduce footprint of React Native Android public APIs we are reducing visibility of classes and interfaces that are not meant to be used publicly OR are public but have no usages.
As part of our analysis, which involved looking for usages inside the Meta codebase and code search in OSS, we've detected that this class/interface is public but it's not used from other packages.

bypass-github-export-checks

If you are using this class or interface please comment in this PR and we will restate the public access.

changelog: [Android][Changed] Reducing visibility of ExecutorException

Reviewed By: cortinico

Differential Revision: D49803280

fbshipit-source-id: 140b5bcd41b1a16ed84196021305a4d27f8cc24a